add special kernel for the upsweep phase of multi-block scan
This kernel calculates the final reduction value for each interval in a multi-block scan. Previously, a standard fold1 was used for this purpose, but that requires a commutative operator which we don't have for the case of segmented scans.

Introduces some mess in the compilation phase, because this falls outside the standard set of AST nodes.

required for AccelerateHS/accelerate#44