rename 'runIn' and friends to 'runWith'
This renaming is motivated by the streaming execution model, which has notions of streaming data in and out, so we can no longer use the *In suffix.

This finishes off @robeverest's half-baked commit AccelerateHS/accelerate-cuda@d8a7337266b3274ba36da655e8bac7ba7c1475b4, in which he renamed some functions but not others. For shame rob, for shame...